Remove test_end_to_end_tax_report
It's not a true end to end test and all of those are already tested at unit/test_accounting.py
